requirements.txt
BeautifulSoup==3.2.1
Flask==0.9
Flask-Assets==0.8
Flask-Script==0.6.7
Flask-WTF==0.9.5
Jinja2==2.6
Pillow==2.4.0
PyYAML==3.11
WTForms==1.0.5
Werkzeug==0.8.3
argparse==1.2.1
arrow==0.4.2
cssselect==0.9.1
docopt==0.6.1
flask-mongoengine==0.7.0
goose-extractor==1.0.8
gunicorn==0.17.3
jieba==0.32
lockfile==0.9.1
lxml==3.3.3
mongoengine==0.8.7
nltk==2.0.4
pymongo==2.6.3
python-dateutil==2.2
six==1.6.1
webassets==0.8
wsgiref==0.1.2
yuicompressor==2.4.7
